requirements.txt
absl-py==2.1.0
accelerate==0.33.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
antlr4-python3-runtime==4.9.3
anyio==4.3.0
apricot-select==0.6.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.1.0
blinker==1.8.2
blobfile==2.1.1
cachetools==5.3.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
chem==1.2.0
click==8.1.7
cmake==3.27.2
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.0
crcmod==1.7
cryptography==42.0.7
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
docker-pycreds==0.4.0
dtrx==8.5.3
easydict==1.13
einops==0.8.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.12.2
fire==0.6.0
flake8==7.0.0
Flask==3.0.3
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.4
imageio==2.34.1
importlib-resources==6.0.1
importlib_metadata==7.0.2
iniconfig==2.0.0
ipykernel==6.29.4
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
lazy-object-proxy==1.10.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning-utilities==0.11.5
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.4
lyft-dataset-sdk==0.0.8
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
mmengine==0.10.4
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
nose==1.3.7
notebook==7.1.2
notebook_shim==0.2.4
numba==0.57.1
numpy==1.24.4
nuscenes-devkit==1.1.11
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.9.0.80
ordered-set==4.1.0
oss2==2.17.0
overrides==7.7.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.0.0
piq==0.8.0
platformdirs==4.2.0
plotly==5.22.0
pluggy==1.5.0
plyfile==1.0.3
pooch==1.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyDeprecate==0.3.2
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.0.9
pyquaternion==0.9.9
pytest==8.2.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytorch-lightning==1.0.8
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rdkit==2023.3.2
rdkit-pypi==2022.9.5
referencing==0.34.0
regex==2023.12.25
requests
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.18.0
safetensors==0.4.4
scikit-learn==1.3.0
scipy
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.44.1
setproctitle==1.3.3
Shapely==1.8.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.2.0
timm==0.4.5
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.4
torch==2.1.0
torchaudio==2.1.0
torchdiffeq==0.2.5
torchmetrics==0.11.4
torchvision==0.16.0
tornado==6.4
tqdm==4.65.2
traitlets==5.14.2
transformers==4.44.0
trimesh==4.4.0
triton==2.1.0
types-python-dateutil==2.9.0.20240316
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
yapf==0.40.2
yarl==1.9.4
zipp==3.16.2
-e .